Moved ChangeInfoConsumer to new file
